split out forum tag code into its own resource
wire up new resources
