+++ server/src/bsworker/bsworker.go
@@ -7,7 +7,6 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/openmultiplayer/web/server/src/resources/burgershot"
-	"github.com/openmultiplayer/web/server/src/resources/forum"
 	"github.com/openmultiplayer/web/server/src/resources/forum/category"
 	"github.com/openmultiplayer/web/server/src/resources/forum/thread"
 	"github.com/openmultiplayer/web/server/src/resources/user"
@@ -16,7 +15,6 @@ import (
 type Worker struct {
 	bs       burgershot.Repository
 	users    user.Repository
-	forum    forum.Repository
 	thread   thread.Repository
 	category category.Repository
 }
@@ -25,12 +23,11 @@ func New(
 	lc fx.Lifecycle,
 	bs burgershot.Repository,
 	users user.Repository,
-	forum forum.Repository,
 	thread thread.Repository,
 	category category.Repository,
 ) Worker {
 
-	w := Worker{bs, users, forum, thread, category}
+	w := Worker{bs, users, thread, category}
 
 	wctx, stop := context.WithCancel(context.Background())
 	lc.Append(fx.Hook{
