--- server/src/resources/forum/repository.go
@@ -1,11 +1,0 @@
-package forum
-
-import (
-	"context"
-
-	"github.com/openmultiplayer/web/server/src/resources/forum/post"
-)
-
-type Repository interface {
-	GetTags(ctx context.Context, query string) ([]post.Tag, error)
-}
