+++ server/src/resources/forum/tag/model.go
@@ -1,0 +1,7 @@
+package tag
+
+type Tag struct {
+	ID    string `json:"id"`
+	Name  string `json:"name"`
+	Posts int    `json:"posts"`
+}
