+++ server/src/resources/forum/tag/repository.go
@@ -1,0 +1,9 @@
+package tag
+
+import (
+	"context"
+)
+
+type Repository interface {
+	GetTags(ctx context.Context, query string) ([]Tag, error)
+}
