--- server/src/resources/forum/type.go
@@ -1,1 +1,0 @@
-package forum
